1、存在会员消费，结账时支付方式默认会员支付 2、固定字体缩放，防止界面布局错位

+++ app/src/main/java/com/wyc/cloudapp/data/room/entity/GiftCardSaleOrder.kt
@@ -291,6 +291,13 @@ class GiftCardSaleOrder():ICardPay<GiftCardSaleDetail> {
         time = System.currentTimeMillis() / 1000
         setOrderPayInfo(a, payDetailList)
 
+        if (payInfo.isEmpty()) {
+            MyDialog.toastMessage(
+                CustomApplication.self().getString(R.string.hints_pay_detail_not_empty)
+            )
+            return
+        }
+
         val dao:GiftCardSaleOrderDao = AppDatabase.getInstance().GiftCardSaleOrderDao();
         dao.deleteWithDetails(this, getSaleInfo(), payInfo)
         dao.insertWithDetails(this, getSaleInfo(), payInfo)
